Heart disease prediction using neural network in python
Deep Neural Network is trained using Back Propagation Algorithm considering the signs and symptoms of a person to predict the chance of occurrence of heart disease. The neural network defined in this project is trained using Cleveland dataset with 14 attributes which are the encoded values for the features that play crucial role in predicting the heart disease.
Thus trained network gave 83% accuracy approximately.

Co-Authored-By: saikrishna challa <50859406+challask@users.noreply.github.com>